anyar

+++ app/Controllers/Pengembalian.php
@@ -8,29 +8,27 @@ class Pengembalian extends BaseController
 {
     public function index()
     {
-        $model = new PengembalianModel();
-        $data['pengembalian'] = $model->findAll();
-
+        $db = \Config\Database::connect();
+        $builder = $db->table('pengembalian');
+    
+        $builder->select('
+            pengembalian.*,
+            peminjaman.tanggal_pinjam,
+            users.nama,
+            buku.judul
+        ');
+    
+        $builder->join('peminjaman', 'peminjaman.id_peminjaman = pengembalian.id_peminjaman');
+        $builder->join('users', 'users.id = peminjaman.id');
+        $builder->join('buku', 'buku.id_buku = peminjaman.id_buku');
+    
+        $builder->orderBy('pengembalian.id_pengembalian', 'DESC');
+    
+        $data['pengembalian'] = $builder->get()->getResultArray();
+    
         return view('pengembalian/index', $data);
     }
-
-    public function create()
-    {
-        return view('pengembalian/create');
-    }
-
-    public function store()
-    {
-        $model = new PengembalianModel();
-
-        $model->insert([
-            'id_peminjaman' => $this->request->getPost('id_peminjaman'),
-            'tanggal_dikembalikan' => $this->request->getPost('tanggal_dikembalikan'),
-            'denda' => $this->request->getPost('denda')
-        ]);
-
-        return redirect()->to('/pengembalian');
-    }
+    
 
     public function edit($id)
     {
@@ -55,9 +53,17 @@ class Pengembalian extends BaseController
 
     public function delete($id)
     {
+        // ❌ blok kalau bukan admin
+        if (session()->get('role') != 'admin') {
+            return redirect()->to('/pengembalian')
+                ->with('error', 'Tidak diizinkan');
+        }
+    
         $model = new PengembalianModel();
         $model->delete($id);
-
-        return redirect()->to('/pengembalian');
+    
+        return redirect()->to('/pengembalian')
+            ->with('success', 'Data berhasil dihapus');
     }
+    
 }
